split app package into focused files
