+++ cmd/server/main.go
@@ -3,8 +3,6 @@ package main
 import (
 	"database/sql"
 	"flag"
-	"fmt"
-	"html/template"
 	"log"
 	"net/http"
 	"os"
@@ -16,6 +14,10 @@ import (
 	_ "github.com/mattn/go-sqlite3"
 )
 
+/*
+main configures the application and starts the HTTP server.
+It wires up the database, templates, routes and middleware.
+*/
 func main() {
 	addr := flag.String("addr", ":8080", "http listen address")
 	dataDir := flag.String("data", "./data", "data directory for sqlite")
@@ -47,7 +49,9 @@ func main() {
 
 	a := &app.App{DB: db, Templates: tpls, CookieName: "forum_session", SessionTTL: 7 * 24 * time.Hour}
 
-	// Routes
+	/*
+	   Set up routes for the application.
+	*/
 	http.HandleFunc("/", a.HandleIndex)
 	http.HandleFunc("/register", a.HandleRegister)
 	http.HandleFunc("/login", a.HandleLogin)
@@ -57,91 +61,11 @@ func main() {
 	http.HandleFunc("/comment/new", a.RequireAuth(a.HandleNewComment))
 	http.HandleFunc("/like", a.RequireAuth(a.HandleLike))
 
-	log.Printf("listening on %s", *addr)
-	// Wrap the mux to handle 404/500
+	/*
+	   Wrap the mux with custom error handlers.
+	*/
 	wrappedMux := withCustomErrors(http.DefaultServeMux, a)
 
 	log.Printf("listening on %s", *addr)
 	log.Fatal(http.ListenAndServe(*addr, logRequest(wrappedMux)))
-}
-
-func logRequest(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		start := time.Now()
-		next.ServeHTTP(w, r)
-		fmt.Printf("%s %s %s\n", r.Method, r.URL.Path, time.Since(start))
-	})
-}
-
-func loadTemplates(dir string) (map[string]*template.Template, error) {
-	layout := filepath.Join(dir, "layout.html")
-	pages, err := filepath.Glob(filepath.Join(dir, "*.html"))
-	if err != nil {
-		return nil, err
-	}
-	m := make(map[string]*template.Template)
-	for _, page := range pages {
-		if filepath.Base(page) == "layout.html" {
-			continue
-		}
-		t, err := template.ParseFiles(layout, page)
-		if err != nil {
-			return nil, err
-		}
-		m[filepath.Base(page)] = t
-	}
-	return m, nil
-}
-
-func withCustomErrors(next *http.ServeMux, app *app.App) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		defer func() {
-			if err := recover(); err != nil {
-				w.Header().Set("Content-Type", "text/html; charset=utf-8")
-				w.Header().Set("Cache-Control", "no-store")
-				w.WriteHeader(http.StatusInternalServerError)
-				if tpl, ok := app.Templates["500.html"]; ok {
-					tpl.ExecuteTemplate(w, "500.html", appTemplateData(r, app))
-				} else {
-					http.Error(w, "Internal Server Error", 500)
-				}
-			}
-		}()
-
-		rw := &responseWriter{ResponseWriter: w, statusCode: 200}
-		next.ServeHTTP(rw, r)
-
-		if rw.statusCode == http.StatusNotFound {
-			if tpl, ok := app.Templates["404.html"]; ok {
-				w.Header().Set("Content-Type", "text/html; charset=utf-8")
-				w.Header().Set("Cache-Control", "no-store")
-				tpl.ExecuteTemplate(w, "404.html", appTemplateData(r, app))
-			} else {
-				http.NotFound(w, r)
-			}
-		}
-	})
-}
-
-// Capture status codes
-type responseWriter struct {
-	http.ResponseWriter
-	statusCode int
-}
-
-func (rw *responseWriter) WriteHeader(code int) {
-	rw.statusCode = code
-	rw.ResponseWriter.WriteHeader(code)
-}
-
-// appTemplateData builds the data map like your other render calls
-func appTemplateData(r *http.Request, app *app.App) map[string]any {
-	uid, uname, logged := app.CurrentUser(r)
-	cats, _ := app.AllCategories()
-	return map[string]any{
-		"LoggedIn":   logged,
-		"UserID":     uid,
-		"Username":   uname,
-		"Categories": cats,
-	}
 }
